2023-10-05_01_00_57 git_src_all_commit

+++ sigmoid.cpp
@@ -2,7 +2,7 @@
 
 float CalcSigmoid(float x)
 	{
-	float y = float(1.0)/(float(1.0) + exp(-x));
+	float y = float(1.0f)/(float(1.0f) + expf(-x));
 	return y;
 	}
